backend/requirements_backend.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose
passlib[argon2]
python-multipart
requests
email-validator
python-dotenv
pillow
